Trash and Spam
I put down the groundwork for the trash and spam folders. There is now a physical trash and spam folder within the file directory that when you click on the trash or spam button, currently it moves the "1.txt" file over as I attempted to follow the standard that is within the email logic. Also currently, It doesn't disreguard the text already there. And there is no display for the trash or spam folders yet, as there is no inbox. Expect an update soon whiping the active mail box text clean.