Email Composition
You can now compose emails and 'send' them. The form checks the 'To: ' field for a valid input (Currently, anything composed of a '@' and '.' will do), then checks to make sure that the source and body have text. If everything checks out, a 'mail sent' popup shows up and then the email form restarts.