Event Logic
This is the update Event Logic and it is as follows:
It keeps track of current events and has a public static field in the main program that should allow other classes to see it.
There are even Broadcaster for all of the current UI component that have been down and are currently being made as well. 
Can be upgraded for future elements as well.