Update
This should be the final implementation update. There will be a small one following this one up with comments and documentation.

I think the timer now counts down between the displayed expiration date and the date the form was opened on. I don't think It will update in real time.

I haven't worked out how to test the times yet, as the ransomware itself is not implemented yet.