Task Manager
It does load processes but on the local computer but I am having trouble making it specific to our project.  Whenever I also try to get the time for each process the application crashes so I need to talk to cherry for that as well.  Still having trouble with icon connecting as well.