Current Task Manager
Current version of Task Manager that opens from the icon and start menu.  the Process class constructor has name be initialized to whichever open form is up in the application and gets the string representation.  However, for some reason when it is booted up nothing appears in it although logically it is sound and I am currently figuring out how to make things appear in there.